Move filtering test out of fixtures package
