+++ subprojects/testing-jvm/src/integTest/groovy/org/gradle/testing/AbstractTestFilteringIntegrationTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013 the original author or authors.
+ * Copyright 2023 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,9 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.testing.fixture
+package org.gradle.testing
 
 import org.gradle.integtests.fixtures.DefaultTestExecutionResult
+import org.gradle.testing.fixture.AbstractJUnitMultiVersionIntegrationTest
 import spock.lang.Issue
 
 abstract class AbstractTestFilteringIntegrationTest extends AbstractJUnitMultiVersionIntegrationTest {
