+++ subprojects/testing-jvm/src/integTest/groovy/org/gradle/testing/junit/AbstractJUnit4FilteringIntegrationTest.groovy
@@ -18,7 +18,7 @@
 package org.gradle.testing.junit
 import org.gradle.integtests.fixtures.DefaultTestExecutionResult
 import org.gradle.integtests.fixtures.TargetCoverage
-import org.gradle.testing.fixture.AbstractTestFilteringIntegrationTest
+import org.gradle.testing.AbstractTestFilteringIntegrationTest
 import spock.lang.Issue
 
 import static org.gradle.testing.fixture.JUnitCoverage.*
